+++ src/Sulu/Bundle/ContactBundle/Contact/ContactManager.php
@@ -279,8 +279,11 @@ class ContactManager extends AbstractContactManager implements DataProviderRepos
 
         $birthday = $this->getProperty($data, 'birthday');
         if (!empty($birthday)) {
-            $contact->setBirthday(new DateTime($birthday));
+            $birthday = new DateTime($birthday);
+        } else {
+            $birthday = null;
         }
+        $contact->setBirthday($birthday);
 
         if (!$id) {
             $parentData = $this->getProperty($data, 'account');