+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-develop
     * BUGIFX      #1871 [ContentBundle]   Fixed url-generation and save button
+    * BUGFIX      #1873 [ContactBundle]   Fixed remove contact birthday
 
 * 1.1.2 (2015-12-11)
     * HOTFIX      #1831 [MediaBundle]     Fixed query for retrieving entities to index